+++ Changelog.md
@@ -23,6 +23,7 @@
 - Refactored Criterion subclasses to remove redundant code
 - Converted "Rename Group" functionality to React modal (#7673)
 - Fixed Rack deprecation warnings by updating HTTP status code symbols (#7675)
+- Refactored "Reuse Groups" functionality to use React modal and relocated button to action box row (#7688)
 
 ## [v2.8.1]